Initial exposition of initializer
